+++ inst/doc/rtfreporter-pipes.R
@@ -57,11 +57,11 @@ knitr::opts_chunk$set(
 # rtf_document() %>%
 #   rtf_tables(list(df1, df2),
 #     titles = list(
-#       c("Table 14.1.1", "{HALF_BLANK_ROW}", "Safety Population"),
+#       c("Table 14.1.1", "", "Safety Population"),
 #       "Table 14.1.2"
 #     ),
 #     footnotes = list(
-#       c("Note 1: foo.", "{HALF_BLANK_ROW}", "Note 2: bar."),
+#       c("Note 1: foo.", "", "Note 2: bar."),
 #       NULL
 #     )
 #   )
@@ -71,13 +71,6 @@ knitr::opts_chunk$set(
 #   rtf_tables(list(df1, df2)) %>%
 #   rtf_titles(list("Page 1", "Page 2")) %>%
 #   rtf_footnotes(list(NULL, "Footer for page 2."))
-
-## ----eval = FALSE-------------------------------------------------------------
-# rtf_header(rows = list(
-#   c(l = "Protocol", r = "Page X"),
-#   c(c = "{HALF_BLANK_ROW}"),       # half-height blank row
-#   c(c = "Table 14.1.1")
-# ))
 
 ## ----eval = FALSE-------------------------------------------------------------
 # doc <- rtf_document() %>%